rngd: Clean up silly command line arguments
rngd has this silly verbose  cmd line option, which prints the available
rng entropy sources and exits.  Thats just not what verbose means.  Lets
remove that flag, and replace it with --list, which seems a bit more
reasonable.  -v is now a synonym for --version, which prints the version
string of the program and exits, which seems much more sane

Note: This is a user interface breaking change.  Anyone using rngd in a
script with this option will need to update said script

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>