rngd: Add nistbeacon support
NIST has a randomness beacon available here:
https://www.nist.gov/programs-projects/nist-randomness-beacon

It generates entropy at a rate of 512 bits per minute.  Its sent in cleartext
over the internet, making it unsuitable for cryptographic function, it is useful
in the generation of entropy for things like monte carlo tests or other uses
where shared pools of entropy might be useful.  As such, lets add the NIST
beacon as an entropy source, but disable it by default so users have to know to
keep it enabled.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>