Fix the default '--fill-watermark' value in the man page
According to rngd man page, the default value for the
'--fill-watermark' option is 2048 bits. However, upon examining
default_watermark() function in rngd_linux.c, it can be seen that this
is true for systems where poolsize cannot be obtained. In linux, the
pool size can be obtained by reading
'/proc/sys/kernel/random/poolsize'. If rngd is capable of reading the
proc entry, as it should be, then it sets the --fill-watermark value
to be 75% of the total pool size, typically 3072 bits.

Bug: https://bugs.gentoo.org/555094