Fix Makefile.am to append libraries rather than prepend them
gcc in ubuntu tahr automatically adds the -as-needed linker flag, which
means that any libraries linked prior to the specification of
application object files will loose symbols that are potentially
required for the application.  Its a wierd behavior, as it spits errors
about undefined symbols that are clearly defined when you check with
objdump.  The solution (shy of changing the compiler behavior), is to
use LD_ADD in the makefile rather than LD_LIBS so that the needed
libraries are linked after the application obj files rather than before.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>