rngd_nistbeacon: fix the size of frequency and timestamp on x86_32
The members "frequency" and "timestamp" of struct nist_data_block are
defined as int and long, respectively. On x86_64, their sizes
correctly correspond to 4 and 8 bytes. However, on 32-bit x86
architectures, both int and long are defined as 4 bytes, causing the
digest verification to fail. Fix it by using uint32_t and uint64_t
explicitly.