configure: modify how we detect the presence of the jitterentropy lib
git submodules store empty directories that need to be populated with a
git submodule update command, meaning using a directory test to enable
its building can lead to false positives and build errors.  Convert the
autoconf test to detect the presence of jitterentropy_library/Makefile
instead to ensure some buildable source is there.  While we're at it,
add in a notice to inform the user in the event that the library isn't
present

Signed-off-by:Neil Horman <nhorman@tuxdriver.com>