Add support for detecting installed libjitterentropy.so
Currently, rng-tools uses libjitterentropy built locally in this source
tree as a static library.  That is sub-optimal however, as jitterentropy
is an independently maintained project.  Ideally it would be nice to
instead load libjitterentropy as a DSO, which that project can build as.
This is the first of three steps toward moving in that direction:

1) Add support to check for and link against libjitterentropy.so, if
found on the host

2) Deprecate local building of jitterentropy as a static library

3) Remove support for building jitterentropy as a static library

This commit implements step 1.  If the configure script finds a local
copy of the jitterentropy source, it will build a static library from
it.  If not, then it will look for a local libjitterentropy.so.  If
found, it will link against that.  Otherwise, it will disable the
jitterentropy source

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>