rngd_jitter: fix process termination on SIGINT and clean mutex locking
It was reported recently that rngd would not terminate properly when
running the jitter entropy source in daemon mode.  It apeared that the
daemon would not terminate on receipt of a signint.  It turns out that
it would eventually terminate, but it would take an extra long time,
because the jitterentropy library was sometimes taking a while to
compute entropy (detecting lots of stuck cycles for some reason).

Fix it by allowing the threads to generate entropy with the mutex
unlocked, and only lock it to update the shared buffer.

There still seems to be a periodic problem with jitterentropy taking
extra long to generate output, so I've added some debug code here as
well to detect when thats happening, in the hopes I can track that down.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>