Add retry count and retry delay options to jitter entropy
jitterentropy can take several seconds to generate entropy (1-2 by my
measurements on my systems, but it could be much longer).  Rngd typicaly
doesn't block waiting for an entropy source, it just marks a lack of
entropy as a failure and moves on. If jitterentropy is your only
entropy source though, too many failures can lead to rngd disabling the
source and potentially shutting down.  When rngd needs entropy, it will
spin quickly, and the errors mount fast.  To avoid that we introduce
these parameters in which, when all jitterentropy buffers are empty,
retry a configurable number of times, sleep a configureable number of
seconds between each retry.  This should allow the buffers to refill.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>