rngd: Link against jitterentropy as a shared library
Using jitterentropy as a git submodule was a fast way to integrate the
project, but really isn't a long term strategy.  Its all the headache of
keeping up with a static library and making releases is a problem to
boot.  Instead, since it builds as a DSO, just link against that if its
installed

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>