JITTER: Support the use of AES to backfill missing entropy
Like with other sources, if JITTER is producing entropy too slowly, we
can add the option to produce pseudo random data using AES encryption.
If a buffer is empty, and this option is enabled, on the first retry
pass, a still-empty buffer can be refilled by encrypting the data with a
random AES key.  Once the JITTER generator has new entropy built, it
will replace the data block with that new true random source data

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>