jitter: fix use of aes
commit aa300c8e5b4d62dcbbadaf91b58b4baaad74c3be elegantly implemented a
ring buffer for our entropy threads, but failed to take into account
the possibility that another task (like the read task when executing
gcrypt operations), might fill the buffer, reducing the needed amount to
refill during the generation cycle.  As a result, the refill tasks were
computing a very large index into the per thread buffer array (->idx is
a size_t / unsinged long), and so we were corrupting memory.  Lets fix
that

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>