Simpify refill code
Theres no need to make the refill path overly complex.  Determining
weather or not to bias for filling or draining not a worthwhile
endeavour.  Once jitter drops the per-thread lock the read side will
have ample time to do all the draining it needs, and if the fill side
can't keep up, our only recourse is to use something like AES to pseudo
randomize anyway.  So if we have more to fill, just go back and fill it.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>