Grab the sleep timer unilaterally
The sleep timer for use in retrying entropy collection when we aren't
using AES wasn't always getting picked up on the first iteration through
(right when we increment retry_count to one), and so we were sleeping on
a garbage value.  Fix it by always grabbing the sleep timer at the start
of a loop iteration after we grab the lock.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>