Mark server_running as volatile
it was pointed out to me that its common practice to mark variables
shared between signal handlers and other code as volatile to prevent the
compiler from optimizing code paths to ignore changes to said variable
that occur in the signal handler.  While I find it hard to believe that
a compiler could not recognize the ability to modify a variable in an
alternate context like a signal handler, they were correct, and all the
C99 and C11 rationale documents support doing this.  As such, mark
server_running as volatile to prevent erroneous optimiation of the
server_running variable

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>