Allow use of --disable-jitterentropy/--enable-jitterentropy
Another attempt to combine the ability to specify both weather or not to
enable jitterentropy on the configure command line and to optionally
specify its location.  With this new macro, jitterentropy is default
enabled, but can be disabled with the --disable-jitterentropy variant.
It can also be expressly enabled with --enable-jitterentropy, and an
option argument provided specifying the path to the libarary location.
If provided, -L <path> and -I <path> are added to LDFLAGS and CFLAGS
respectively so that the library can be found

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>