Fix assemby textrels on rdrand_asm.S on PIC x86
This patch updates the fixes in the assembly in rdrand_asm.S in
sys-apps/rng-tools-5 so it won't generate textrels on PIC systems.
The main fixes are in the use of leal in SETPTR for such systems, the rest is
the usual PIC support stuff.

This should fix Gentoo bug #469962 and help fix #518210

This patch is released under the GPLv2 or a higher version license as is the
original file as long as the author and the tester are credited.

Gentoo-bug-url: https://bugs.gentoo.org/show_bug.cgi?id=469962
Gentoo-bug-url: https://bugs.gentoo.org/show_bug.cgi?id=518210
Signed-off-by: Francisco Blas Izquierdo Riera (klondike) <klondike@gentoo.org>
Reported-by: cilly <cilly@cilly.mine.nu>
Reported-by: Manuel Rüger <mrueg@gentoo.org>
Tested-by: Anthony Basile <blueness@gentoo.org>