rngd_jitter: Convert implementation to use a pipe to feed xread_jitter
Based on suggestions from sgrubb@redhat.com, convert jitter entropy
source to use a pipe for collection of entropy.  This simplifies the
thread communication model, in that signaling of threads to wake up and
generate more entropy is no longer needed.  They will simply block on
writing to the pipe when its full, and wake up when there is space to
add more entropy.

Suggested-by: Steve Grubb <sgrubb@redhat.com>
Signed-off-by: Neil Horman <nhorman@redhat.com>