rngd: Fix watermark sizing
just found a huge bug in the default watermark computation, which was causing
our watermark to always default to 2048 bits, which is half the size of the
kernel pool.  This was preventing us from waking up early enough to refill the
entropy pool when its getting drained rapidly.

While we're at it, lets also record the pools maximum size and use that to
ensure we keep filling until its topped off completely.  This hysteresis makes
sure that rngd is agressive in filling the entropy pool completely whenever it
starts draining

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>