Use RNDGETENTCNT to gate sleeping in rngd
Recently introduced changes to create hysteresis in waking and sleeping
of rngd were not completed.  Specifically testing for the size of the
entropy pool was broken as it was assumed the RNDADDENTCNT ioctl
returned the resultant fill size of the entropy pool.  We actually need
to use RNDGETENTCNT to retrieve that number, allowing rngd to go back to
sleep when the pool is full

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>