Remove SIGTERM/SIGINT handling
The previous code had bugs, see:
62fbff0ac807e07f4f21f58c0f15bfe5e9c8b55e
The version in RHEL doesn't have this patch, and we're hitting
shutdown bugs in RHEL CoreOS.

There's no actual good reason though to have a `SIGTERM` handler
at all; doing things like `pthread_join()` is pointless becuase
the kernel will take care of removing all resources.

In fact it's really an anti-pattern to do this, because if every
process has these types of things, it makes shutdowns much less
efficient as the process needs to be fully scheduled in rather
than just having the kernel take care of it.