Daemonize earlier in init
It was reported here:
https://github.com/nhorman/rng-tools/issues/73

That rngd hangs on shutdown, but only when being run as a daemon.  The problem
turns out to be that we are daemonizing too late in init.  Because we daemonize
after initalizing all our entropy sources, we get a premature exit on the
threads created prior to daemonization.  This leads to hangs in shutdown when we
wait for termination of threads that have already exited abnormally.

Fix it by daemonizing before we init any entropy sources

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>