Enhance message_entsrc to work when rngd runs as a daemon
when running as a deemon, rngd logs to syslog, which inserts newlines on every
call to syslog.  As such we need to log every message as a single syslog call to
avoid unwanted newlines.  So modify message_entsrc to merge both the source
prefix and the message to a single buffer before printing

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>