Add reserved option value to not change write_wakeup_threshold
several distributions use a separate single shot service to set the
write_wakeup_threashold, which allows for non-root users to run rngd.  This
results in a warning however, when rngd attempts to write the
write_wakeup_threshold.  Avoid this by allowing a reserved value, 0 for the
fill-watermark option that causes rngd to just not touch the
write_wakeup_threshold

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>