rngd.c: only print the success message when initialize
Only print the message for the entropy sources which
initialized successfully to avoid misleading users.

That makes sense as after check the available entropy
sources, there is logic to check whether there is at
least one available entropy source exist as below:
if (!ent_sources) {
     message(LOG_DAEMON|LOG_ERR,
             "can't open any entropy source");
     message(LOG_DAEMON|LOG_ERR,
             "Maybe RNG device modules are not loaded\n");
     return 1;
}

That's to say, only printing the message for the entropy
sources which initialized successfully is safe.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>