Add forced reseeding of kernel entropy pool
With newer kernels containing non blocking entropy pools, it is still
beneficial to add external entropy on a periodic basis to guard entropy
prediction.  Add a --force_reseed option to rngd which defines a period
on which rngd will write additional entropy into the kernel, even if its
entropy pool is reading as full.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>