Fix RNDADDENTROPY ioctl error for lrng
The Linux Random Number Generator (lrng) returns the number of bytes as
the result of the RNDADDENTROPY ioctl instead of 0. This causes
rng-tools to report an error like this:

  RNDADDENTROPY failed: Success

After so many failures, rng-tools will exit and stop adding entropy and
exit.

A positive ioctl return value is not indicative of an explicit error and
limiting the range of error return values to negative values will fix
this issue and allow rng-tools to support kernels using either crng or
lrng.