Rewrite init_kernel_rng() to ensure proper logging
We want to log errors in default_watermark(), so actually we should not
call it before logging is initialized, i.e. before processing command
line switches and setting am_daemon. With that, default_watermark() is
less needed. Adjust the code by merging it into init_kernel_rng() which
is called exactly after logging is initialized and "-l" case is handled.

Also use LOG_DEBUG priority for informational messages about entropy pool.
Initialize arguments->fill_watermark with -1 so we can distinguish cases
when watermark was or was not set by a command line switch.

Signed-off-by: Vladis Dronov <vdronov@redhat.com>