Prevent jitter from operating on a single cpu with software timers
We recently added the ability to run software timers in jitterentropy.
Thats good.  But the software timers don't help anything if theres only
a single core, and the free running counter is too coarse to deliver
timing information.  Thats bad.  If that situation arises (as it
apparently does with older single core athlon processors) just don't
allow jitterentropy to run.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>