Fix an ugly potential null deref
Vladis pointed out here:
https://github.com/nhorman/rng-tools/issues/141

That we had a potential null deref in nistbeacon, that wasn't easily
fixed, and thei is the best way I can come up with to address it

Basically, when we parse a nist json block returned from a server, we
need to be aware that some memory allocations that occur in the process
might fail. If they do, we should abort the curl_easy_perform operation
so that the calling fuction knows that the block is invalid. To do that
I have to check the return code of every element parser we have (which
we use several times).  Thats alot of extra if(rc){return 0} code, so
I'm codifying it in a variadic macro that returns if the call fails

yes, I'm returning in a macro. Yes, thats ugly.  And yes, there is a
huge comment warning all who enter that code that chaos and madness lie
within.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>