Fix up version option handling
The man page was out of date, specifying -V as the man page version
flag, when It should be -v.  Also, the arg parser wasn't handling those
options well, as it was returning an error for those valid options.  Fix
that up to just exit cleanly after printing the version info