Build Power9 code separately
The PPC64 DARN instruction is defined in ISA 3.0 which is for Power9 CPUs.
Power8 CPUs are ISA 2.07 and thus do not support DARN. Recently gcc has
limited the DARN instruction to power9 target only. Build a code with DARN
in a separate library with special compile flags.

Also add a missing check for get_darn() return value as in other places.