Adjust jitterentropy library to timeout/fail on long delay
When running rngd -l its possible, on platforms that have low jitter
entropy to block for long periods of time.  Adjust jitter to timeout on
init after 5 seconds in the event it takes to long to gether jitter
entropy

Also while we're at it, I might have a build solution for the presence
of internal timers.  When jitterentropy is built without internal
timers, jent_notime_init is defined publically, but when it is built
with timers, its declared as a static symbol, preenting resolution, so
we can test to see if the function exists.  If it does we _don't_ have
notime support. The logic is a bit backwards, but i think it works