add option force_as_fast_source to jitter source
By default the jitter entropy source is considered a "slow" source, meaning it's output is only used
when no "fast" sources are available or able to provide sufficient random data. When setting this option
to 1 jitter will be considered as a "fast" source too and it's output will always be considered (default 0).

Having this option is useful when you want to feed the kernel data from a mix of as many independent sources
as possible. The concern of jitter being too slow to supply a sufficient amount of entropy data is only
relevant when rngd needs to supply entropy to the kernel often and in short order. With modern kernels
not signaling a low watermark anymore and thus only the known force-reseed time is used, you can ensure
that jitter will be able to provide enough data at configuration time.

When comparing the default timeout of 5 seconds for the jitter source with the default force-reseed time
of 5 minutes, you can see that jitter will often be able to provide enough data.