Add error handling to jitter read and pipe write loops
Add a counter for a maximum number of errors. If it is reached, bail
out. Reset the counter on successful read or write.

Add error handling to the jitter read loop. Restart a read on error,
since there is nothing to write out.

Add error handling to the pipe write loop. Handle a special case of
EBADF error, see 3e47faa for details.

Fixes #209.